move network duplication code/testing into separate files
