+++ src/MLJBase.jl
@@ -202,6 +202,7 @@ include("composition/abstract_types.jl")
 include("composition/learning_networks/nodes.jl")
 include("composition/learning_networks/inspection.jl")
 include("composition/learning_networks/machines.jl")
+include("composition/learning_networks/replace.jl")
 
 include("composition/models/methods.jl")
 include("composition/models/from_network.jl")
